pyproject.toml
[project]
name = "pybmds"
license = {text = "MIT License"}
description = "A Python interface for the USEPA Benchmark dose modeling software (BMDS)"
authors = [
  {name = "BMDS development team"},
]
readme = "README.md"
version = "24.1a0"
classifiers = [
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
]
requires-python = ">=3.11"
dependencies = [
  "matplotlib>=3.7",
  "numpy",
  "openpyxl",
  "pandas>=2",
  "pydantic>=2",
  "python-docx",
  "scipy",
  "tabulate",
  "tqdm",
]

[project.optional-dependencies]
dev = [
  "coverage~=7.5.1",
  "mypy~=1.10.0",
  "pybind11~=2.10.4",
  "pytest-mpl~=0.17.0",
  "pytest~=8.2.1",
  "ruff~=0.4.4",
  "setuptools~=67.7.2",
  "wheel~=0.43.0",
]
docs = [
  "Sphinx~=7.3.7",
  "myst-nb~=1.1.0",
  "myst-parser~=3.0.1",
  "pydata-sphinx-theme~=0.15.2",
  "sphinx-autobuild~=2024.4.16",
]

[project.scripts]
"bmds-priors-report" = "bmds.cli.priors_report:main"

[build-system]
requires = [
  "pybind11>=2.10.4",
  "setuptools>=67.7.2",
]
build-backend = "setuptools.build_meta"

[tool.cibuildwheel]
test-requires = "pytest"
test-command = "pytest {project}/tests"

[tool.coverage.run]
include = [
  "src/pybmds/*"
]

[tool.pytest.ini_options]
testpaths = "tests/test_pybmds"
addopts = "--mpl-baseline-path=tests/data/mpl"
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["F", "E", "W", "I", "UP", "S", "B", "T20", "RUF"]
lint.ignore = ["E501", "B904", "B007", "S308", "S113", "S314"]
lint.isort.known-first-party = ["pybmds"]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["S101", "S106"]
"setup.py" = ["S603", "S607"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["pybmds*"]
